requirements/local.txt
-r base.txt

django-extensions
black
django-debug-toolbar
